Make repo management optional
this introduced a `manage_repo` parameter. By default we will install
the puppet-tools release rpm or configure the yumrepo. When
`manage_repo` is false, we won't do any of that and assume that the user
has a repo that provides puppet-bolt. Useful for Satellite/Katello
users or when you want to use distro repos.